Course Class Remaining Assignment Percent Calculator Added
Created a function to check the current total score against a desired end score percentage (i.e. A is 90%, b is 80%, etc) and returns the percentage required for the remaining assignments within the course (which will always add up 1000 total points). This percentage can be used with the new Assignment calculateGradeFromPercentage() function in order to determine what exact point grades are needed from each future assignment in the system in order to achieve the desired grade.

Changelog:
- Added calculatePercentNeededForGrade() into Course class