Delete Latest Assignment Functionality Added
Created the removeAssignmentLastAdded() function to Course class. This function will delete the assignment object out of the ArrayList that was last added to it. Currently, this is called using the Import Grade button, but that button will be renamed in the future.
Also added some error checking to getAssignmentLastAdded() to ensure a correct error is thrown if the function is called when no assignments have been added to the database.