Course Class now checks if Assignment Name already exists in ArrayList
Added a check within Course Class method addAssignment that throws an IllegalArgumentException if an assignment is passed in that shares a name with an assignment that already exists. This exception is displayed in a JOptionPane pop up.

KNOWN ISSUE: Because of how the AssignmentDisplay text field is updating during the Controller Class actionPerformed method, the duplicate object is still displayed. However, this is a ghost object that doesn't actually exist. This issue will be fixed when frontend devs refactor the updateAssignmentDisplay method and remove display code from actionPerformed.