File Save/Load Bug Fixes and Automatic .dat Extension
Updated the JFileChooser method within Controller class to resolve several errors. Also changed the method signature to JFileChooser(String chooserType) so that the Save window would have a"Save" button and the Open window an "Open" one. Lastly, updated the file output on the save execution to add the .dat file extension automatically.